pyproject.toml
[tool.black]
line-length=80

[tool.poetry]
name = "blog-platform"
version = "0.1.0"
description = ""
authors = ["macalistervadim <macalistervadim@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
python-dotenv = "^1.0.1"
django-debug-toolbar = "^4.4.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-ckeditor = "^6.7.1"
django-taggit = "^5.0.1"
pillow = "^10.4.0"
django-taggit-serializer = "^0.1.7"
django-cleanup = "^8.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "7.0.0"
flake8-absolute-import = "^1.0.0.2"
flake8-builtins = "^2.5.0"
flake8-bugbear = "^24.4.26"
flake8-clean-block = "^0.1.2"
flake8-commas = "^4.0.0"
flake8-comments = "^0.1.2"
flake8-comprehensions = "^3.14.0"
flake8-eradicate = "^1.5.0"
flake8-fixme = "^1.1.1"
flake8-import-order = "^0.18.2"
flake8-implicit-str-concat = "^0.4.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
flake8-return = "^1.2.0"
flake8-type-ignore = "^0.1.0.post2"
flake8-use-pathlib = "^0.3.0"
pep8-naming = "0.13.3"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"